Move printing and parsing of types to a new file.
