+++ src/mlir/dialect/VeronaTypes.cc
@@ -340,261 +340,6 @@ namespace mlir::verona
     return getImpl()->right;
   }
 
-  /// Parse a list of types, surrounded by angle brackets and separated by
-  /// commas. The types inside the list must be Verona types and should not use
-  /// the `!verona.` prefix.
-  ///
-  /// Empty lists are allowed, but must still use angle brackets, i.e. `< >`.
-  /// Lists of one elements are also allowed.
-  static ParseResult
-  parseTypeList(DialectAsmParser& parser, llvm::SmallVectorImpl<Type>& result)
-  {
-    if (parser.parseLess())
-      return failure();
-
-    if (succeeded(parser.parseOptionalGreater()))
-      return success();
-
-    do
-    {
-      mlir::Type element = parseVeronaType(parser);
-      if (!element)
-        return failure();
-
-      result.push_back(element);
-    } while (succeeded(parser.parseOptionalComma()));
-
-    if (parser.parseGreater())
-      return failure();
-
-    return success();
-  }
-
-  static Type parseMeetType(MLIRContext* ctx, DialectAsmParser& parser)
-  {
-    SmallVector<mlir::Type, 2> elements;
-    if (parseTypeList(parser, elements))
-      return Type();
-    return MeetType::get(ctx, elements);
-  }
-
-  static Type parseJoinType(MLIRContext* ctx, DialectAsmParser& parser)
-  {
-    SmallVector<mlir::Type, 2> elements;
-    if (parseTypeList(parser, elements))
-      return Type();
-    return JoinType::get(ctx, elements);
-  }
-
-  static Type parseIntegerType(
-    MLIRContext* ctx, DialectAsmParser& parser, StringRef keyword)
-  {
-    size_t width = 0;
-    if (keyword.substr(1).getAsInteger(10, width))
-    {
-      parser.emitError(parser.getNameLoc(), "unknown verona type: ") << keyword;
-      return Type();
-    }
-    bool sign = keyword.startswith("S");
-    return IntegerType::get(ctx, width, sign);
-  }
-
-  // Annoyingly, DialectAsmParser only exposes `parseOptionalString`, no
-  // `parseString`. This method implements the latter based on the former.
-  ParseResult parseString(DialectAsmParser& parser, StringRef* value)
-  {
-    auto loc = parser.getCurrentLocation();
-    if (failed(parser.parseOptionalString(value)))
-      return parser.emitError(loc, "expected string literal");
-    else
-      return success();
-  }
-
-  static Type parseClassType(MLIRContext* ctx, DialectAsmParser& parser)
-  {
-    auto loc = parser.getNameLoc();
-
-    StringRef name;
-    SmallVector<std::pair<StringRef, Type>, 4> fields;
-
-    if (parser.parseLess() || parseString(parser, &name))
-      return Type();
-
-    // TODO: Support parsing recursive types by constructing an uninitialized
-    // ClassType first and filling it up later. This would require passing
-    // around the set of "pending" classes around.
-    // See the LLVM dialect's implementation of struct types for an example.
-    while (succeeded(parser.parseOptionalComma()))
-    {
-      StringRef field_name;
-      Type field_type;
-      if (
-        parseString(parser, &field_name) || parser.parseColon() ||
-        !(field_type = parseVeronaType(parser)))
-        return Type();
-
-      fields.push_back({field_name, field_type});
-    }
-
-    if (parser.parseGreater())
-      return Type();
-
-    ClassType pending = ClassType::get(ctx, name);
-    if (failed(pending.setFields(fields)))
-    {
-      InFlightDiagnostic diag = parser.emitError(loc)
-        << "class type \"" << pending.getName()
-        << "\" already used with different definition";
-      return Type();
-    }
-
-    return pending;
-  }
-
-  static Type parseViewpointType(MLIRContext* ctx, DialectAsmParser& parser)
-  {
-    Type left;
-    Type right;
-    if (
-      parser.parseLess() || !(left = parseVeronaType(parser)) ||
-      parser.parseComma() || !(right = parseVeronaType(parser)) ||
-      parser.parseGreater())
-      return Type();
-
-    return ViewpointType::get(ctx, left, right);
-  }
-
-  static Type
-  parseFloatType(MLIRContext* ctx, DialectAsmParser& parser, StringRef keyword)
-  {
-    size_t width = 0;
-    if (keyword.substr(1).getAsInteger(10, width))
-    {
-      parser.emitError(parser.getNameLoc(), "unknown verona type: ") << keyword;
-      return Type();
-    }
-    return FloatType::get(ctx, width);
-  }
-
-  Type parseVeronaType(DialectAsmParser& parser)
-  {
-    MLIRContext* ctx = parser.getBuilder().getContext();
-
-    StringRef keyword;
-    if (parser.parseKeyword(&keyword))
-      return Type();
-
-    if (keyword == "class")
-      return parseClassType(ctx, parser);
-    else if (keyword == "meet")
-      return parseMeetType(ctx, parser);
-    else if (keyword == "join")
-      return parseJoinType(ctx, parser);
-    else if (keyword == "viewpoint")
-      return parseViewpointType(ctx, parser);
-    else if (keyword == "top")
-      return MeetType::get(ctx, {});
-    else if (keyword == "bottom")
-      return JoinType::get(ctx, {});
-    else if (keyword == "iso")
-      return CapabilityType::get(ctx, Capability::Isolated);
-    else if (keyword == "mut")
-      return CapabilityType::get(ctx, Capability::Mutable);
-    else if (keyword == "imm")
-      return CapabilityType::get(ctx, Capability::Immutable);
-    else if (keyword == "bool")
-      return BoolType::get(ctx);
-    else if (keyword.startswith("U") || keyword.startswith("S"))
-      return parseIntegerType(ctx, parser, keyword);
-    else if (keyword.startswith("F"))
-      return parseFloatType(ctx, parser, keyword);
-
-    parser.emitError(parser.getNameLoc(), "unknown verona type: ") << keyword;
-    return Type();
-  }
-
-  static void printTypeList(ArrayRef<Type> types, DialectAsmPrinter& os)
-  {
-    os << "<";
-    llvm::interleaveComma(
-      types, os, [&](auto element) { printVeronaType(element, os); });
-    os << ">";
-  }
-
-  void printClassType(ClassType type, DialectAsmPrinter& os)
-  {
-    // TODO: Support printing recursive types. If we're already in the process
-    // of printing a class' body and reach the same class again, we should skip
-    // its body. See the LLVM dialect's implementation of struct types for an
-    // example.
-    os << "class<\"" << type.getName() << "\"";
-    for (auto [field_name, field_type] : type.getFields())
-    {
-      os << ", \"" << field_name << "\" : ";
-      printVeronaType(field_type, os);
-    }
-    os << ">";
-  }
-
-  void printVeronaType(Type type, DialectAsmPrinter& os)
-  {
-    TypeSwitch<Type>(type)
-      .Case<IntegerType>([&](IntegerType type) {
-        if (type.getSign())
-        {
-          os << "S";
-        }
-        else
-        {
-          os << "U";
-        }
-        os << type.getWidth();
-      })
-      .Case<FloatType>([&](FloatType type) { os << "F" << type.getWidth(); })
-      .Case<BoolType>([&](BoolType type) { os << "bool"; })
-      .Case<MeetType>([&](MeetType type) {
-        if (type.getElements().empty())
-        {
-          os << "top";
-        }
-        else
-        {
-          os << "meet";
-          printTypeList(type.getElements(), os);
-        }
-      })
-      .Case<JoinType>([&](JoinType type) {
-        if (type.getElements().empty())
-        {
-          os << "bottom";
-        }
-        else
-        {
-          os << "join";
-          printTypeList(type.getElements(), os);
-        }
-      })
-      .Case<CapabilityType>([&](CapabilityType type) {
-        switch (type.getCapability())
-        {
-          case Capability::Isolated:
-            os << "iso";
-            break;
-          case Capability::Mutable:
-            os << "mut";
-            break;
-          case Capability::Immutable:
-            os << "imm";
-            break;
-        }
-      })
-      .Case<ClassType>([&](ClassType type) { printClassType(type, os); })
-      .Case<ViewpointType>([&](ViewpointType type) {
-        os << "viewpoint<" << type.getLeftType() << ", " << type.getRightType()
-           << ">";
-      });
-  }
-
   bool isaVeronaType(Type type)
   {
     return type.isa<
